removed redundant check for key === null
