+++ lib/context/index.js
@@ -1,114 +1,3 @@
-'use strict'
-/**
- * @function createContextGetter
- * @memberOf Base#
- * @param  {string} key - Key to create the context getter for
- */
-exports.createContextGetter = function (key) {
-  const value = this.contextCandidate(this[key])
-  if (value) {
-    let privateKey = '_' + key
-    let keys = value._keys
-    if (keys) {
-      for (let i = 0, len = keys.length; i < len; i++) {
-        console.log(keys[i])
-        value.createContextGetter(keys[i])
-      }
-    }
-    if (!(privateKey in this)) {
-      this[privateKey] = value
-      Object.defineProperty(this, key, {
-        get () {
-          const value = this[privateKey]
-          if (value) {
-            if (!this.hasOwnProperty(privateKey)) {
-              value.__c = this
-              value._cLevel = 1
-            } else if (this.__c) {
-              value._cLevel = this._cLevel + 1
-              value.__c = this.__c
-            } else {
-              value.clearContext()
-            }
-          }
-          return value
-        },
-        set (val) {
-          // can handle null here
-          this[privateKey] = val
-        },
-        configurable: true
-      })
-    }
-  }
-}
-
-/**
- * @function resolveContext
- * resolves context on set
- * creates instances up to the point where a set within context is performed
- * @memberOf Base#
- * @param {*} val set value to be resolved
- * @param {stamp} stamp current stamp
- * @param {base} context resolve this context
- * @param {boolean} alwaysreturn when set to true always returns resolved base
- * @type {base|undefined}
- */
-exports.resolveContext = function (val, stamp, context, target, path) {
-  context = context || this.__c
-  if (!target) { target = this }
-  var i = this._cLevel
-  var prevContext
-  var level = i
-  var iterator = target
-
-  if (!path) {
-    path = pathMaker(path, iterator, i)
-  }
-
-  if (context.__c) {
-    let cpath = pathMaker(false, context, context._cLevel)
-    context = context.resolveContext(val, stamp, context.__c, target, cpath.concat(path))
-    return context
-  }
-
-  let pathLength = path.length
-  let pathLengthCompare = pathLength - 1
-
-  for (i = 0; i < pathLength; i++) {
-    if (context) {
-      prevContext = context
-      let segment = path[i]
-      let key = typeof segment === 'string' ? segment : segment.key
-      let prop = context[key] || segment.__cTarget__
-      let set
-      if (i === pathLengthCompare) {
-        set = val
-      } else {
-        set = {}
-      }
-      // double check if not having the stamp here is correct need to add more tests
-      context = context.setKeyInternal(key, set, false, true, prop)
-      if (!context) {
-        context = prevContext[key]
-      }
-    }
-  }
-  this.clearContextUp(level)
-  return context
-}
-
-function pathMaker (path, iterator, i) {
-  path = path || []
-  while (i) {
-    let key = iterator.key
-    if (iterator._cKey) {
-      path.unshift({ key: key, __cTarget__: iterator })
-    } else {
-      path.unshift(key)
-    }
-    iterator = iterator._parent
-    i--
-  }
-  return path
-}
+'user strict'
+exports.createContextGetter = require('./getter')
+exports.resolveContext = require('./resolve')
