+++ lib/remove.js
@@ -11,9 +11,6 @@ const vstamp = require('vigour-stamp')
  * @return {boolean|undefined} if true nothing happened
  */
 exports.removeUpdateParent = function (parent, stamp, context) {
-  if (this.key === null) {
-    return
-  }
   if (parent[this.key] === null) {
     return true
   }
